Fix 13 compile errors from CI build
- AudioAnalyzer: replace nonisolated methods with Task.detached to avoid
  cross-isolation access to gain/smoothingAlpha properties
- VideoExporter: fix MainActor.run closure type inference for CVPixelBuffer
- CartoonFilter: explicit Float cast for CIPixellate scale
- VisualEffectsView: use Color.accentColor instead of .accentColor
- WaveformView: wrap ternary branches in AnyShapeStyle for type matching
- EyeAnimatorView: fix CGImage non-optional binding, Float.ulpOfOne type
- MouthAnimator: fix CGImage non-optional binding from texture.cgImage()
- StickerPackExporter: replace UIImage.draw(in:from:) with CGImage cropping

+++ PetTalk/Services/CartoonFilter.swift
@@ -122,7 +122,7 @@ enum CartoonFilter {
         case .pixellate:
             let pixellate = CIFilter.pixellate()
             pixellate.inputImage = input
-            pixellate.scale = max(input.extent.width / 80, 4)
+            pixellate.scale = Float(max(input.extent.width / 80, 4))
             return pixellate.outputImage
 
         case .crystallize:
